requirements.txt
torch
torchvision
streamlit
pydicom
opencv-python-headless
pillow
numpy
pandas
pytorch-grad-cam
